refactor LittleBabyMan.java to CommandProcessor.java
Moved command processing from LittleBabyMan.java to CommandProcessor.java
